[update]마이페이지 회원정보 수정 이미지 업로드

+++ src/main/java/com/sparta/picboy/service/myPage/MyPageService.java
@@ -343,24 +343,28 @@ public class MyPageService {
     public ResponseDto<?> updateimage(UserDetails userinfo, MultipartFile file) {
         Member member = memberRepository.findByUsername(userinfo.getUsername()).orElse(null);
         if (member == null) return ResponseDto.fail(ErrorCode.NOT_FOUND_MEMBER);
-        String thisImgUrl = member.getProfileImg();
-        String deleteUrl = thisImgUrl.substring(thisImgUrl.indexOf("picboy/mypageImg"));
-
-        System.out.println(deleteUrl);
-        //deleteUrl : picboy/mypageImg/4354c745-097f-4230-8f32-03e2347ae3d2-mg
-        awsS3Service.deleteImage(deleteUrl);
-        //amazonS3Client.deleteObject(new DeleteObjectRequest(bucket, fileName));
-        String imageUrl = getFileUrl(file);
-        if (imageUrl == null) return ResponseDto.fail(ErrorCode.FAIL_FILE_UPLOAD);
-        member.updateImg(imageUrl);
-        memberRepository.save(member);
+        if (member.getProfileImg().equals(null)) {
+            String imageUrl = getFileUrl(file);
+            if (imageUrl == null) return ResponseDto.fail(ErrorCode.FAIL_FILE_UPLOAD);
+            member.updateImg(imageUrl);
+            memberRepository.save(member);
+        } else {
+            String thisImgUrl = member.getProfileImg();
+            String deleteUrl = thisImgUrl.substring(thisImgUrl.indexOf("picboy/mypageImg"));
+            System.out.println(deleteUrl);
+            //deleteUrl : picboy/mypageImg/4354c745-097f-4230-8f32-03e2347ae3d2-mg
+            awsS3Service.deleteImage(deleteUrl);
+            //amazonS3Client.deleteObject(new DeleteObjectRequest(bucket, fileName));
+            String imageUrl = getFileUrl(file);
+            if (imageUrl == null) return ResponseDto.fail(ErrorCode.FAIL_FILE_UPLOAD);
+            member.updateImg(imageUrl);
+            memberRepository.save(member);
+        }
 
         //substring 예제
         //String finds = "abc-def";
         //String ans =  finds.substring(0,3); //abc
         //String ans2 =  finds.substring(4); //def
-
-
 
         return ResponseDto.success("이미지를 수정하였습니다.");
     }
